src/main/java/ru/job4j/ood/srp/reports/ReportHR.java
package ru.job4j.ood.srp.reports;

import java.util.Comparator;
import java.util.List;
import java.util.function.Predicate;

public class ReportHR implements Report {

    private final Store store;

    public ReportHR(Store store) {
        this.store = store;
    }

    @Override
    public String generate(Predicate<Employee> filter) {
        StringBuilder text = new StringBuilder();
        text.append("Name; Salary;").append(System.lineSeparator());
        List<Employee> salarySort = store.findBy(filter);
        salarySort.sort(Comparator.comparingDouble(Employee::getSalary));
        for (Employee employee : salarySort) {
            text.append(employee.getName()).append(" ")
                    .append(employee.getSalary())
                    .append(System.lineSeparator());
        }
        return text.toString();
    }
}
